+++ app/src/main/java/com/example/radiologi/LoginActivity.java
@@ -3,13 +3,11 @@ package com.example.radiologi;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.android.volley.Request;
@@ -18,6 +16,8 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
+import com.example.radiologi.home.DataAdmin;
+import com.example.radiologi.home.DataDokter;
 
 import org.json.JSONArray;
 import org.json.JSONException;
