+++ app/src/main/java/com/example/radiologi/home/DataDokter.java
@@ -1,4 +1,4 @@
-package com.example.radiologi;
+package com.example.radiologi.home;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
@@ -18,6 +18,10 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
+import com.example.radiologi.DataDokterPagerAdapter;
+import com.example.radiologi.LoginActivity;
+import com.example.radiologi.R;
+import com.example.radiologi.SharedPreferenceManager;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.android.material.tabs.TabLayout;
