+++ app/src/main/res/layout/activity_data_admin.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".DataAdmin">
+    tools:context=".home.DataAdmin">
 
     <RelativeLayout
         android:id="@+id/bar"
